+++ src/main.cpp
@@ -52,7 +52,7 @@ PYBIND11_MODULE(_core, m) {
                                       [](py::object) { return ProjectComparator::ByCostDescThenVotesDesc; });
 
     py::class_<Election>(m, "Election")
-        .def(py::init<long long, long long, std::vector<ProjectEmbedding>>(), "budget"_a, "num_voters"_a, "projects"_a)
+        .def(py::init<long long, int, std::vector<ProjectEmbedding>>(), "budget"_a, "num_voters"_a, "projects"_a)
         .def_property_readonly("budget", &Election::budget)
         .def_property_readonly("num_voters", &Election::numVoters)
         .def_property_readonly("projects", &Election::projects);
